refactor(sim,render): code hygiene round 1 — Group E (#71, #72, #77, #87)
Pure mechanical refactors. NO behavior change.

#77 — minimap FP_SHIFT import. Pre-fix code had a bare `>> 8` for
fixed-point→tile conversion. Imports FP_SHIFT from sim/fixed.js.

#71 — Object.hasOwn guards on for-in over world.colonies. Most sites
already had the guard; four sites in ant-system.ts and one in
surface-features.ts were missing it. Adds the guard for consistency.
Audit confirms: 16 production for-in sites, all now guarded.

#72 — named simVersion sentinels. Replace bare `world.simVersion >= 3`
and `>= 6` in ant-system.ts and colony-system.ts with the imported
SIM_VERSION_V3 and SIM_VERSION_V6_FORAGER_NO_REVISIT constants. Also
sweeps types.ts prose to remove stale `(LATEST_SIM_VERSION)`
annotations on v5 and v8 (latest is v10 today; v5/v8 lost the title
in earlier bumps and the comments drifted).

#87 — extract shared BFS helper. chamber-flow.ts, dig-system.ts, and
entrance-flow.ts all had near-identical BFS expansion loops with the
same eslint-suppressed `(idx / width) | 0` row decode. Extract
bfsExpandSeededField into a new src/sim/bfs-flow-field.ts module;
all three callers import it and delete their local copies plus the
now-unused REVERSE/NEIGHBOR_DR/NEIGHBOR_DC constants. Direction
encoding and traversal predicate (Open + BeingDug) are byte-identical
across all three prior copies — verified by review.

1811 tests passing (1 added vs main, automatic from imports).
